requirements.txt
botocore==1.34.82
boto3==1.34.82
psycopg2-binary